build with vs

+++ recipes/mpfr/all/conanfile.py
@@ -1,5 +1,6 @@
-from conans import ConanFile, AutoToolsBuildEnvironment, tools
+from conans import ConanFile, AutoToolsBuildEnvironment, CMake, tools
 import os
+import shutil
 
 
 class MpfrConan(ConanFile):
@@ -13,10 +14,13 @@ class MpfrConan(ConanFile):
     settings = "os", "arch", "compiler", "build_type"
     options = {"shared": [True, False], "fPIC": [True, False]}
     default_options = {"shared": False, "fPIC": True}
-    requires = "gmp/6.1.2"
+    requires = "mpir/3.0.0"
+    exports_sources = "CMakeLists.txt"
+    generators = "cmake"
 
     _source_subfolder = "source_subfolder"
     _autotools = None
+    _cmake = None
 
     def config_options(self):
         if self.settings.os == 'Windows':
@@ -30,6 +34,9 @@ class MpfrConan(ConanFile):
         tools.get(**self.conan_data["sources"][self.version])
         extracted_dir = self.name + "-" + self.version
         os.rename(extracted_dir, self._source_subfolder)
+        if self.settings.compiler == "Visual Studio":
+            shutil.copy("CMakeLists.txt",
+                        os.path.join(self._source_subfolder, "CMakeLists.txt"))
 
     def _configure_autotools(self):
         if not self._autotools:
@@ -48,19 +55,33 @@ class MpfrConan(ConanFile):
             self._autotools.configure(args=args, configure_dir=self._source_subfolder)
         return self._autotools
 
+    def _configure_cmake(self):
+        if not self._cmake:
+            self._cmake = CMake(self)
+            self._cmake.configure(source_folder=self._source_subfolder)
+        return self._cmake
+
     def build(self):
-        autotools = self._configure_autotools()
-        autotools.make(args=["V=0"])
+        if self.settings.compiler == "Visual Studio":
+            cmake = self._configure_cmake()
+            cmake.build()
+        else:
+            autotools = self._configure_autotools()
+            autotools.make(args=["V=0"])
 
     def package(self):
         self.copy("COPYING", dst="licenses", src=self._source_subfolder)
-        autotools = self._configure_autotools()
-        autotools.install()
-        la = os.path.join(self.package_folder, "lib", "libmpfr.la")
-        if os.path.isfile(la):
-            os.unlink(la)
-        tools.rmdir(os.path.join(self.package_folder, "share"))
-        tools.rmdir(os.path.join(self.package_folder, "lib", "pkgconfig"))
+        if self.settings.compiler == "Visual Studio":
+            cmake = self._configure_cmake()
+            cmake.install()
+        else:
+            autotools = self._configure_autotools()
+            autotools.install()
+            la = os.path.join(self.package_folder, "lib", "libmpfr.la")
+            if os.path.isfile(la):
+                os.unlink(la)
+            tools.rmdir(os.path.join(self.package_folder, "share"))
+            tools.rmdir(os.path.join(self.package_folder, "lib", "pkgconfig"))
 
     def package_info(self):
         self.cpp_info.libs = ["mpfr"]
